web3.py : added support

+++ contracts/1. SimpleStorage.sol
@@ -2,24 +2,25 @@
 pragma solidity ^0.6.0;
 
 contract SimpleStorage {
-    uint num;
-    
+    uint256 num;
+
     struct People {
-        uint num;
+        uint256 num;
         string name;
     }
-    
+
     People[] public arr;
-    mapping (string => uint) hashmap;
+    mapping(string => uint256) hashmap;
+
     function store(uint256 _num) public {
         num = _num;
     }
-    
-    function retrive() public view returns (uint256){
+
+    function retrive() public view returns (uint256) {
         return num;
     }
-    
-    function addPerson(uint _num, string memory _name) public {
+
+    function addPerson(uint256 _num, string memory _name) public {
         arr.push(People(_num, _name));
         hashmap[_name] = _num;
     }
